Revert "[compat] fixed warning"
This reverts commit 7444102948186c61feaca3597e2144a443556dd0.

+++ src/lib/rtm/CorbaNaming.cpp
@@ -1035,6 +1035,6 @@ namespace RTC
         begin_pos = found_pos + delim_size;
         pre_pos   = found_pos + delim_size;
       }
-    return (unsigned int)results.size();
+    return results.size();
   }
 };  // namespace RTC
